Windows tested successful

+++ pdm-crypt-win/cc20_multi.h
@@ -29,7 +29,7 @@ class Cc20 {
 public:
 
     void start_seq();
-    void encr(uint8_t* line, uint8_t* linew, unsigned long int fsize);
+    void encr(uint8_t* line, uint8_t* linew, unsigned long long int fsize);
     void rd_file_encr(const std::string file_name, std::string oufile_name);
     void stream(uint8_t* plain, unsigned int len);
     void set_vals(uint8_t* nonce, uint8_t* key);
